+++ test/test-server/src/thirdparty.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import ThirdParty from "../../../recipe/thirdparty";
-import { handleSession, serializeRecipeUserId, serializeUser } from "./utils";
+import { convertRequestSessionToSessionObject, serializeRecipeUserId, serializeUser } from "./utils";
 import { logger } from "./logger";
 
 const namespace = "com.supertokens:node-test-server:thirdparty";
@@ -9,7 +9,7 @@ const { logDebugMessage } = logger(namespace);
 const router = Router().post("/manuallycreateorupdateuser", async (req, res, next) => {
     try {
         logDebugMessage("ThirdParty:manuallyCreateOrUpdateUser %j", req.body);
-        let session = req.body.session && (await handleSession(req.body.session));
+        let session = req.body.session && (await convertRequestSessionToSessionObject(req.body.session));
         const response = await ThirdParty.manuallyCreateOrUpdateUser(
             req.body.tenantId || "public",
             req.body.thirdPartyId,